% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_ovule_parameters.R
\name{curve_ovule_parameters}
\alias{curve_ovule_parameters}
\title{Estimation of max. and min. numbers of visits for full fertilization}
\usage{
curve_ovule_parameters(Crop, self = 0)
}
\arguments{
\item{Crop}{Crop species (botanical name) to be evaluated}

\item{self}{number of pollen grains present on the stigmas before first bee visit to account for selfing (default = 0)

External parameters used in the function:

svd = Single Visit Pollen Deposition of different bee species on a certain crop obtained from data measured in the field 
ovu = mean number of ovules of the inspected crop
eff_min = (ovu-self)/(svd + ovu - self) minimum efficiency to achieve full fertilization
target_visits = target number of visit for full fertilization. It is obtained by using eff_min
min_visits = minimum number of visit for full fertilization. It is obtained by using eff = 1}
}
\value{
table with curve parameters to calculate ovule fertilization curves for different crop/pollinator/pollination efficiency scenarios (also to be used in crop_parameter_table)
}
\description{
Function to estimate the max. and min. number of visits to full fertilization.
It depends on "evolution_ovules.R" function, used in the crop_parameter_table.R function
}
\examples{
data_crop <- curve_ovule_parameters("Blueberry",0)

i=14 # Visitor: Colletes sp. 

svd <- data_crop$svd[i]
ovu <- data_crop$ovu[i]
self <- data_crop$self[i]
example_1 <- tibble(visits=c(0:60),
                   perc_fecundated_ovules = evolution_ovules(svd,ovu,self,1,c(0:60)),
                   efficiency = paste0("eff. = 1.00"))

ggplot(example_1, aes(x=visits,y=perc_fecundated_ovules,color =efficiency))+
geom_point()+
geom_line()+
labs(x="Visits",y="Percentage of fecundated ovules",
     color="Fertilization\n Efficiency",
     title = data_crop$Species[i])
     
}
