% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_visitation_rates.R
\name{eval_visitation_rate}
\alias{eval_visitation_rate}
\title{Plot of yield curves}
\usage{
eval_visitation_rate(Crop, visit_rate)
}
\arguments{
\item{Crop}{non-latin name of the crop (e.g. "Apple", "Blueberry", etc.)}

\item{visit_rate}{visitation rate measured in the center of a given field}
}
\value{
Plot showing the observed yield and visitation rates as well as the max. yield and the visitation rates necessary to obtain it  
works if the Crop is in the referred tabl otherwise an error message will be produced
}
\description{
Function that creates yield curves based on the data in the synthetic_table 
To generate the yield curve we use the lower target visitation rate in the previous table.
}
\examples{
eval_visitation_rate(Crop ="Blueberry",11)

}
