% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_inflicted_yield.R
\name{non_inflicted_yield}
\alias{non_inflicted_yield}
\title{Non-inflected curve function}
\usage{
non_inflicted_yield(a, b, c, visit_rate)
}
\arguments{
\item{a}{obtained percentage of max. yield without pollinators present}

\item{b}{Max. yield attained for the crop}

\item{c}{Shape of the curve}

\item{visit_rate}{observed visitation rate (visits in 100 flowers during 1 hour)}
}
\value{
Yield curve function for increasing number of pollinators used in the plot_yield_curves
}
\description{
Creates a non inflected curve function (Exponential decay - increasing form) for the plot_yield_curves function
}
