% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_par_exp_decay.R
\name{shape_par_exp_decay}
\alias{shape_par_exp_decay}
\title{Shape parameter (c) for an exponential decay (increasing form)}
\usage{
shape_par_exp_decay(a, b, visit_rate_full_fer, threshold = 99)
}
\arguments{
\item{a}{percentage of yield without pollinators}

\item{b}{Max. yield attained for the crop}

\item{visit_rate_full_fer}{visits in 100 flowers during 1 hour for full fertilization (Garibaldi 2019)}

\item{threshold}{percentage of yield that is considered full fertilization (standard value = 99)}
}
\value{
parameter c for non_inflicted_yield function
}
\description{
Creates the shape parameter c for the non_inflected yield function
}
