% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_meta_data-data.R
\docType{data}
\name{study_meta_data}
\alias{study_meta_data}
\title{Felix Bross compilation of studies with data on pollination}
\format{
An object of class \code{"tibble"}.
}
\usage{
data(study_meta_data)
}
\description{
Data from a references on crop pollination
}
\keyword{datasets}
