% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_garibaldi2019-data.R
\docType{data}
\name{table_garibaldi2019}
\alias{table_garibaldi2019}
\title{Table 1 of Garibaldi et al. (2020)}
\format{
An object of class \code{"tibble"}.
}
\source{
\href{https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/1365-2664.13574}{Link}
}
\usage{
data(table_garibaldi2019)
}
\description{
Data table 1 of Garibaldi et al. (2020)
}
\references{
Garibaldi, L. A., Sáez, A., Aizen, M. A., Fijen, T., & Bartomeus, I. (2020). Crop pollination management needs flower‐visitor monitoring and target values. Journal of Applied Ecology, 57(4), 664–670.
(\href{https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/1365-2664.13574}{Link})
}
\keyword{datasets}
