% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits_table_syn-data.R
\docType{data}
\name{visits_table_syn}
\alias{visits_table_syn}
\title{Felix Bross compilation of data on crop pollination}
\format{
An object of class \code{"tibble"}.
}
\usage{
data(visits_table_syn)
}
\description{
Data on crop pollination from several references, with information on 
the number of ovules per flower for a given crop, and an estimation of single
pollen deposition for the studied floral visitors.
}
\keyword{datasets}
