% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_parameter_table.R
\name{yield_parameter_table}
\alias{yield_parameter_table}
\title{Creates a table with yield_parameters to be used in the "model_eval_visitation_rate.R" function}
\usage{
yield_parameter_table()
}
\value{
returns a table with parameters to be used in "model_eval_visitation_rate.R" function
}
\description{
Create yield_parameter_table that gives selfing values and max. production values for different crops
}
\references{
check .txt file
}
