% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_variables-data.R
\docType{data}
\name{yield_variables}
\alias{yield_variables}
\title{Felix Bross compilation of data on crop yields}
\format{
An object of class \code{"tibble"}.
}
\usage{
data(yield_variables)
}
\description{
Data on crop yields from several references, with information on 
the maximum production for a given crop, and its production without insect
visitors.
}
\keyword{datasets}
